+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 0.5.0-alpha.1
+
+Released 2026-Jul-01
+
 * Add support for reporting remote configuration status.
   ([#4509](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4509))